Create README.md
RMS allows the users to perform their assigned tasks and customers to make reservations about their preferred areas. This database contains information about the employees, the customers, the tables, and all the services that are offered. It can be used by employees in a restaurant to handle the clients, their orders and can help them easily find free tables